% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconErrSymPmf.R
\name{DeconErrSymPmf}
\alias{DeconErrSymPmf}
\title{Estimate Discrete Deconvolution of Data With Symmetric Error}
\usage{
DeconErrSymPmf(W, m = 10, n.iter.tp = 5, n.iter.var = 2,
  show.diagnostics = F)
}
\arguments{
\item{W}{A vector of the contaminated data}

\item{m}{The number of support points to use in finding the Pmf}

\item{n.iter.tp}{The number of times to attempt minimization of T(p)}

\item{n.iter.var}{The number of times to attempt minimization of the variance}

\item{show.diagnostics}{Flag indicating whether to show diagnostics}
}
\value{
A list with components:
	\item{support}{The support of the deconvolved distribution}
	\item{probweights}{The sizes of the probability mass at each respective 
					   point of support.}
	\item{phi.W}{The empirical distribution of W and the t values on which it is 
				 calculated}
}
\description{
Estimates deconvolution of data \eqn{W = X + U}, where \eqn{U} is symmetric, 
with a probability mass function.
}
\details{
Details here
}
\examples{
# Generate Test Data
W <- GenerateTestData(n = 500)

# Deconvolve
out <- DeconErrSymPmf(W)

phi.W <- out$phi.W

# Convert to Probability Density
fX <- DeconErrSymPmfToPdf(out, W, phi.W)

# Plot results
PlotPmfAndPdf(out, fX)
}
