% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve}
\alias{deconvolve}
\title{Deconvolution}
\usage{
deconvolve(W, xx, errortype = NULL, sigU = NULL, phiU = NULL, h = NULL,
  rescale = FALSE, phiK = NULL, muK2 = 6, RK = 1024/3003/pi,
  tt = seq(-1, 1, 2e-04))
}
\description{
Description
}
\details{
Details
}
