% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconvolveSymmetricErrorPmf.R
\name{DeconvolveSymmetricErrorPmf}
\alias{DeconvolveSymmetricErrorPmf}
\title{Estimate Discrete Deconvolution of Data With Symmetric Error}
\usage{
DeconvolveSymmetricErrorPmf(W)
}
\arguments{
\item{W}{A vector of the contaminated data}
}
\value{
A list with components:
	\item{support}{The support of the deconvolved distribution}
	\item{probweights}{The sizes of the probability mass at each respective 
					   point of support.}
}
\description{
Estimates deconvolution of data \eqn{W = X + U}, where \eqn{U} is symmetric, 
with a probability mass function.
}
\details{
Details here
}
