% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framingham.R
\docType{data}
\name{framingham}
\alias{framingham}
\title{Framingham Data}
\format{A data frame with 1615 rows and 10 variables:
\describe{
\item{OBS}{observation number (1-1615)}
	\item{AGE}{age at exam 2}
	\item{SBP21}{first systolic blood pressure at exam 2}
	\item{SBP22}{second systolic blood pressure at exam 2}
	\item{SBP31}{first systolic blood pressure at exam 3}
	\item{SBP32}{second systolic blood pressure at exam 3}
	\item{SMOKE}{present smoking at exam 1}
	\item{CHOLEST2}{serum cholesterol at exam 2}
	\item{CHOLEST3}{serum cholesterol at exam 3}
	\item{FIRSTCHD}{indicator of first evidence of CHD occurring at exam 3 through 6, i.e., within an eight-year follow-up period to exam 2 }
}}
\source{
\url{https://www.stat.tamu.edu/~carroll/data.php}
}
\usage{
framingham
}
\description{
A dataset from the Framingham Study on coronary heart disease described by 
Carroll, Ruppert, and Stefanski (1995)
}
\section{References}{
 
Carroll, R. J., Ruppert, D., and Stefanski, L. A. (1995). Measurement Error 
in Nonlinear Models: A Modern Perspective, Second Edition. Chapman Hall, New
York.
}

\keyword{datasets}
