% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateTestData.R
\name{GenerateTestData}
\alias{GenerateTestData}
\title{Generate convolved data for use in deconvolve examples}
\usage{
GenerateTestData(n, sig_X = 1, sig_U = 0.2, dist_type = "chi",
  error_type = "norm")
}
\arguments{
\item{n}{The size of the generated data}

\item{sig_X}{The standard deviation of X}

\item{sig_U}{The standard deviation(s) of U. For heteroscedastic errors,
supply \code{sig_U} as a length n vector.}

\item{dist_type}{One of \code{'chi'} or \code{'mix'}. The distribution type
of X.}

\item{error_type}{One of \code{'norm'} or \code{'Lap'}. The distribution type
of U.}
}
\value{
Returns W as a length n vector.
}
\description{
A convenience function for the package `deconvolve' that generates 
contaminated data of the form \eqn{W = X + U} which can be used to test 
the function \code{deconvolve}.
}
\author{
Aurore Delaigle, Timothy Hyndman, Tianying Wang
}
