% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_deconvolve.R
\name{reg_deconvolve}
\alias{reg_deconvolve}
\title{Compute the measurement error version of the Nadaraya-Watson regression
estimator}
\usage{
reg_deconvolve(xx, W, Y, errortype, sigU, h = NULL, rho = NULL,
  n_cores = 2)
}
\arguments{
\item{xx}{vector of x-values where to compute the regression estimator}

\item{W}{vector of contaminated data W_1, ..., W_n}

\item{Y}{vector of data Y_1, ..., Y_n}

\item{errortype}{'Lap' for Laplace errors and 'norm' for normal errors.}

\item{sigU}{parameter of Laplace or normal errors used only to define characteristic function of the error.}

\item{h}{bandwidth}

\item{rho}{ridge parameter. See Delaigle,  A. and Hall,  P. (2008). Using SIMEX for smoothing-parameter choice
in errors-in-variables problems.  JASA,  103,  280-287.}

\item{n_cores}{cores used to do parallel computation to calculate bandwidth.}
}
\value{
Regression estimator.
}
\description{
Goal: estimate m where Y=m(X)+epsilon,  and we observe data on (W, Y),  where
W=X+U.
See Fan,  J.,  and Truong,  Y. K. (1993),  Nonparametric Regression With
Errors in Variables,  The Annals of Statistics,  21,  1900-1925
}
\section{Warnings}{

\itemize{
\item h and rho need to be consistent. They need to be both specified or unspecified.
\item The range of t-values -1 and 1 correspond to the support of phiK.
\item If the grid of t-values is fine enough,  the estimator can simply be
computed like here without having problems with oscillations.
\item The phiK here is also used to compute the bandwidth.
\item The estimator can also be computed using the Fast Fourier Transform,  which
is faster,  but more complex.
\item See Delaigle,  A. and Gijbels,  I. (2007). Frequent problems in calculating
integrals and optimizing objective functions: a case study in density
deconvolution.   Statistics and Computing,   17,   349 - 355.}
}

\section{References}{

Fan,  J.,  and Truong,  Y. K. (1993),  Nonparametric Regression With Errors in Variables,  \emph{The
Annals of Statistics}.  21,  1900-1925.

Delaigle,  A. and Hall,  P. (2008). Using SIMEX for smoothing-parameter choice in errors-in-variables
problems.  \emph{JASA},  103,  280-287.
}

\examples{
\dontrun{
	n <- 200
	sigX <- 1
	sigU <- 0.2
	data <- GenerateTestData(n, sigX, sigU, dist_type = "mix", error_type = "norm", create_Y = TRUE)
	xx <- seq(-2, 8, 0.1)
	output <- deconvolve::reg_deconvolve(xx = xx, W = data$W, Y = data$Y, errortype = "norm", sigU = 0.2)
}
}
