% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R
\name{bandwidth}
\alias{bandwidth}
\title{Bandwidth Selectors for Deconvolution Kernel Density Estimation}
\usage{
bandwidth(W, errortype, sigU, phiU, varX = NULL, algorithm = "PI",
  phiK = NULL, muK2 = 6, RK = 1024/3003/pi, tt = seq(-1, 1, 2e-04))
}
\description{
Description
}
\details{
The function \code{bandwidth} chooses between three different algorithms
to calculate a bandwidth depending on the inputs.
}
\section{Problems}{

}

