% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconErrSymPmf.R
\name{DeconErrSymPmf}
\alias{DeconErrSymPmf}
\title{Deconvolution when the error distribution is symmetric}
\usage{
DeconErrSymPmf(W, m = 10, n.iter.tp = 5, n.iter.var = 2,
  show.diagnostics = F)
}
\arguments{
\item{W}{A vector of the contaminated data}

\item{m}{The number of support points to use in finding the Pmf}

\item{n.iter.tp}{The number of times to attempt minimization of T(p)}

\item{n.iter.var}{The number of times to attempt minimization of the variance}

\item{show.diagnostics}{Flag indicating whether to show diagnostics}
}
\value{
A list with components:
	\item{support}{The support of the deconvolved distribution}
	\item{probweights}{The sizes of the probability mass at each respective 
					   point of support.}
	\item{phi.W}{The empirical distribution of W and the t values on which it is 
				 calculated}
}
\description{
Estimates the distribution of \eqn{X} with a probability mass function from 
data \eqn{W = X + U}, when the distribution of \eqn{U} is symmetric.
}
\details{
PUT DETAILS HERE
}
\section{Authors}{

Aurore Delaigle, Timothy Hyndman
}

\section{References}{

Delaigle, A. and Hall, P. (2016). Methodology for non-parametric 
deconvolution when the error distribution is unknown. \emph{Journal of the 
Royal Statistical Society: Series B (Statistical Methodology)}, 78, 1, 
231-252.
}

\examples{
# Generate Test Data
W <- GenerateTestData(n = 500)

# Deconvolve
out <- DeconErrSymPmf(W)

phi.W <- out$phi.W

# Convert to Probability Density
xx <- seq(min(W), max(W), length.out = 100)
fX <- DeconErrSymPmfToPdf(out, W, phi.W, xx)
}
