% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconErrSymPmfToPdf.r
\name{DeconErrSymPmfToPdf}
\alias{DeconErrSymPmfToPdf}
\title{Create Probability Density Function from Probability Mass Function}
\usage{
DeconErrSymPmfToPdf(X.pmf, W, phi.W)
}
\arguments{
\item{X.pmf}{A list with parts "support" and "probweights"}

\item{W}{A vector of the contaminated data}

\item{phi.W}{A list with parts "complex", "re", "im", "norm" and "t.values" 
containing phi.W, Re(phi.W), Im(phi.W), Norm(phi.W) and the t values on which
they were calculated respectively.}
}
\value{
A list with components:
\item{x}{The values on which \eqn{f_X} was calculated}
\item{y}{The density \eqn{f_X(x)} evaluated at each x}
}
\description{
Takes the output of DeconErrSymPmf along with the contaminated data, W, and
uses a kernel estimator to find a probability density function for X.
}
\details{
Details here
}
\examples{
# Generate Test Data
W <- GenerateTestData(n = 500)

# Deconvolve
out <- DeconErrSymPmf(W)

phi.W <- out$phi.W

# Convert to Probability Density
fX <- DeconErrSymPmfToPdf(out, W, phi.W)

# Plot results
PlotPmfAndPdf(out, fX)
}
