% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\docType{package}
\name{deconvolve}
\alias{deconvolve}
\alias{deconvolve-package}
\title{Computing the deconvolution kernel density estimator and its 
bandwidths in the heteroscedastic case}
\description{
This package illustrates how to use the functions for computing the 
deconvolution kernel density estimator and its bandwidths in the 
heteroscedastic case.
}
\details{
\tabular{ll}{
Package: \tab deconvolve\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-03-18\cr
License: \tab  GPL>=2\cr
}
}
\references{
Delaigle, A. and Meister, A. (2008). Density estimation with
heteroscedastic error. Bernoulli, 14, 562-579
}
\author{
Aurore Delaigle

Maintainer: Tianying Wang <tianying@stat.tamu.edu>
}
\keyword{package}
