% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateTestData.r
\name{GenerateTestData}
\alias{GenerateTestData}
\title{Generate convolved data for use in deconvolve examples}
\usage{
GenerateTestData(n, sd_X = 1, sd_U = 0.2, dist_type = "chi",
  error_type = "norm", create_Y = FALSE)
}
\arguments{
\item{n}{The size of the generated data}

\item{sd_X}{The standard deviation of X}

\item{sd_U}{The standard deviation(s) of U. For heteroscedastic errors,
supply \code{sd_U} as a length n vector.}

\item{dist_type}{One of \code{'chi'} or \code{'mix'}. The distribution type
of X.}

\item{error_type}{One of \code{'norm'} or \code{'Lap'}. The distribution type
of U.}

\item{create_Y}{If true, \code{GenerateTestData} also returns a vector of Y 
values for use in regression problems.}
}
\value{
Returns W as a length n vector. If \code{create_Y} is true, returns a 
list with elements 'W' and 'Y'.
}
\description{
A convenience function for the package `deconvolve' that generates 
contaminated data of the form \eqn{W = X + U} which can be used to test 
the function \code{deconvolve}.
}
\author{
Aurore Delaigle, Timothy Hyndman, Tianying Wang
}
