% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_deconvolve.R
\name{reg_deconvolve}
\alias{reg_deconvolve}
\title{Compute the measurement error version of the Nadaraya-Watson regression
estimator}
\usage{
reg_deconvolve(W, Y, xx, errortype = NULL, sigU = NULL, phiU = NULL,
  bw = NULL, rho = NULL, n_cores = NULL, kernel_type = "default")
}
\arguments{
\item{W}{A vector of the univariate contaminated data W_1, ..., W_n.}

\item{Y}{A vector of the response data Y_1, ..., Y_n.}

\item{xx}{A vector of x values on which to compute the regression estimator.}

\item{errortype}{The distribution type of \eqn{U}. Either "Lap" for Laplace 
errors or "norm" for normal errors.}

\item{sigU}{The standard deviation of \eqn{U}. This does not need to be 
provided if you define your error using phiU and provide \code{bw} and 
\code{rho}.}

\item{phiU}{A function giving the characteristic function of \eqn{U}. If you 
define the errors this way then you should not provide \code{errortype}.}

\item{bw}{The bandwidth to use. If you provide this then you should also
provide \code{rho}.}

\item{rho}{The ridge parameter to use. If you provide this then you should 
also provide \code{bw}.}

\item{n_cores}{Number of cores to use when calculating the bandwidth. If
\code{NULL}, the number of cores to use will be automatically detected.}

\item{kernel_type}{The deconvolution kernel to use. The default kernel has
characteristic function \eqn{(1-t^2)^3}.}
}
\value{
Regression estimator, bandwidth and ridge parameter rho. Using SIMEX 
to choose smoothing-parameters. See Delaigle and Hall 2008.
}
\description{
Estimates \eqn{m} from data \eqn{(W, Y)} where \eqn{Y = m(X) + \epsilon} and 
\eqn{W = X + U}. See Fan and Truong 1993.
}
\section{Warnings}{

\itemize{
\item If provided, the bandwidth \code{h} and ridge parameter \code{rho} need 
to be consistent. You should either provide both or neither.
\item The estimator can also be computed using the Fast Fourier Transform, 
which is faster, but more complex. See Delaigle and Gijbels 2007.
}
}

\section{References}{

Fan,  J.,  and Truong,  Y. K. (1993),  Nonparametric Regression With Errors 
in Variables,  \emph{The Annals of Statistics}.  21,  1900-1925.

Delaigle, A. and Hall, P. (2008). Using SIMEX for smoothing-parameter choice 
in errors-in-variables problems. \emph{JASA},  103,  280-287.

Delaigle, A. and Gijbels, I. (2007). Frequent problems in calculating
integrals and optimizing objective functions: a case study in density
deconvolution. \emph{Statistics and Computing}, 17, 349 - 355.
}

\examples{
n <- 200
sigX <- 1
sigU <- 0.2
data <- GenerateTestData(n, sigX, sigU, dist_type = "mix", error_type = "norm", create_Y = TRUE)
xx <- seq(-2, 8, 0.1)
output <- reg_deconvolve(data$W, data$Y, xx, errortype = "norm", sigU = 0.2, n_cores = 2)
}
